fix #16 fix message encrypt/decrypt based on js lib fix
