+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tsjs-xpx-catapult-sdk",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "description": "Proximax catapult sdk for typescript and javascript",
   "scripts": {
     "pretest": "npm run build",
@@ -45,7 +45,7 @@
     "mocha": "^4.0.1",
     "nyc": "^11.3.0",
     "ts-mockito": "^2.2.7",
-    "ts-node": "^5.0.1",
+    "ts-node": "^8.2.0",
     "tslint": "^5.8.0",
     "typescript": "^3.4.4",
     "typescript-require": "^0.2.9-1"
@@ -54,7 +54,7 @@
     "@types/crypto-js": "^3.1.43",
     "crypto-js": "^3.1.9-1",
     "js-joda": "^1.6.2",
-    "js-xpx-catapult-library": "^0.1.0",
+    "js-xpx-catapult-library": "^0.2.0",
     "request": "^2.83.0",
     "request-promise-native": "^1.0.5",
     "rxjs": "^6.2.1",
